Animation runs smoother now

+++ app/src/main/java/de/andipaetzold/dodgeit/objects/GameObject.java
@@ -1,14 +1,14 @@
 package de.andipaetzold.dodgeit.objects;
 
 public abstract class GameObject {
-    protected int x;
-    protected int y;
+    protected float x;
+    protected float y;
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
